+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - [#398](https://github.com/estruyf/vscode-front-matter/issues/398): Fix Windows folder path parsing in data folder retrieval
 - [#400](https://github.com/estruyf/vscode-front-matter/issues/400): Fix for draft/published content grouping
 - [#403](https://github.com/estruyf/vscode-front-matter/issues/403): Fix for media files with spaces on importing in article content
+- [#404](https://github.com/estruyf/vscode-front-matter/issues/404): Fix for published sorting option in media dashboard
 
 ## [8.0.1] - 2022-07-13